Add pagination functionality to the main page
This commit introduces client-side pagination to the main page. We added a util package to better handle number conversions which are crucial for the introduced features. This resolves user report issues about inefficient loading and interaction on the main page when there is a large amount of metadata. Pagination allows users to navigate between different chunks of metadata, rendering a limited number of items at a time, thus significantly improving the user experience. Furthermore, changes in the metadata categories have been made to match the new structure fetched from the database. Lastly, we have also introduced the Sanity client to handle database operations more efficiently.