Add error handling for 'to' prop and create blog page
Added a check in 'a.tsx' to throw an error when 'to' property is used instead of 'href'. This ensures proper usability of the component and prevent incorrect usage.

Additionally, created a new blog page, '_blog.blog.moving-away-from-vercel.tsx', which includes usage of multiple components such as 'A', 'Blockquote', 'Heading', and 'Paragraph'. This change is necessary to provide a platform for blog posts on the website.

Also created 'blockquote.tsx' for handling blockquotes within the blog posts. This addition improves the readability of quotes in the blogs.