Add new blog post and improve link element
Added a new blog post titled 'Revisiting Old GraphQL Friend'. The post explains the author's recent experience with Apollo GraphQL and NextJS 13 App Router, and their decision to remove GraphQL from their project. Additionally, added the prefetching property to the anchor (a) element to improve performance as this allows Next.js to fetch the required data for the linked page in the background.