Add Rock-Paper-Scissors game to blog post
Added a new feature to the blog post section: an interactive Rock-Paper-Scissors game. The feature includes a game logic along with a layout consisting of buttons representing rock, paper, or scissors and information about win rates, streaks and last round result. This addition will serve to engage the audience in a fun and interactive way. Additionally, new YouTube component has been added to include video content in posts, and aspect ratio options are expanded in the styling config file for better layout management.