"Add new blog post on web storage"
This commit includes the addition of a new blog post file '_blog.blog.web-storage-is-cool.tsx' offering a detailed explanation of web storage options. Adjustments are also made to 'pnpm-lock.yaml', 'package.json', and new file 'storage-estimate.tsx' to add and implement 'convert' dependency which is used in the new blog post for formatting storage values. This change provides readers a clear and comprehensive understanding of different web storage methods.