Add blog post explaining next image component's behavior
Added a new blog post for the purpose of shedding light on the confusion regarding the behavior of new 'Next/Image' component in NextJS. The post explains why width and height are still required by default in the component and how objectFit still needs to be defined in a relative container even when using the fill property. This is intended to clear any misunderstanding and encourage correct usage of the Next/Image component.