Add function for checking environment variables
Added a new function for validating environment variables. This function throws an error if any of the predefined set of environment variables are missing or not defined. This assists in preventing runtime errors due to undeclared or undefined environment variables, which improves application reliability and fault tolerance. This is part of the blog post tutorial explaining how to programmatically ensure essential environment variables are set in a Node.js application.