Rounded scroll progress and simplified aria-label
It was noticed that the scroll progress value wasn't rounded, which made the progress bar visually inaccurate. Therefore, the scrollWidth value is now rounded to the nearest whole digit for an accurate depiction of reading progress to the user. In addition, the aria-label was simplified to only say "Scroll Progress" instead of printing the value in the label, as users using screen readers don't need to hear that verbose information.