Adjust grid column size in course list
The grid column size for the course list was increased from 300px to 400px. This was done by modifying both the CSS class in the course component as well as the corresponding configuration in Tailwind. This increase in size will allow for larger course card components to be displayed without breaking the layout.